上传声强改为数组型, 地址分自动与手动分开上传

+++ NoiseComplain/NCPMeterView.h
@@ -13,8 +13,8 @@
  */
 @interface NCPMeterView : UIView
 
-@property(nonatomic, assign, setter=setValue:) float value;
+@property(nonatomic, assign, setter=setValue:) double value;
 
-- (void)setValueWithLable:(float)value;
+- (void)setValueWithLable:(double)value;
 
 @end
